+++ src/cases/signup/fr_hp_signup.py
@@ -5,8 +5,8 @@ import logging
 from selenium.webdriver.chrome.webdriver import WebDriver
 
 from src.config import Config as cfg
-from src.pages.homepage.fr_homepage import HomepageFR
-from src.pages.application.fr_application import RegistrationFormsFR
+from src.pages.france.homepage import Homepage
+from src.pages.france.application import Application
 from src.reporter import Reporter
 from src.mailbox import Mailbox
 
@@ -20,7 +20,7 @@ class FRHomeSignupChrome(unittest.TestCase):
     def setUp(self) -> None:
         self.name = "FRHomeSignupChrome"
         self.driver = WebDriver(cfg.CHROMEDRIVER_PATH)
-        self.page = HomepageFR(self.driver)
+        self.page = Homepage(self.driver)
 
         self.reporter = Reporter(self.driver, self.page.iso, self.name)
 
@@ -251,7 +251,7 @@ class FRHomeSignupChrome(unittest.TestCase):
         self.reporter.write(step_n, step)
 
         # Landing in registration forms
-        reg_forms = RegistrationFormsFR(self.driver)
+        reg_forms = Application(self.driver)
 
         # Verify regforms loaded
         step_n += 1
