+++ src/pages/__init__.py
@@ -1,0 +1,46 @@
+from .argentina import (
+    Application as ApplicationAR,
+    Homepage as HomepageAR
+)  # noqa
+from .canada import (
+    Application as ApplicationCA,
+    Homepage as HomepageCA
+)  # noqa
+from .france import (
+    Application as ApplicationFR,
+    Homepage as HomepageFR
+)  # noqa
+from .germany import (
+    Application as ApplicationDE,
+    Homepage as HomepageDE
+)  # noqa
+from .russia import (
+    Application as ApplicationSU,
+    Homepage as HomepageSU
+)  # noqa
+from .united_kingdom import (
+    Application as ApplicationUK,
+    Homepage as HomepageUK
+)  # noqa
+from .united_states import (
+    Application as ApplicationUS,
+    Homepage as HomepageUS
+)  # noqa
+
+
+__all__ = [
+    "ApplicationAR",
+    "HomepageAR",
+    "ApplicationCA",
+    "HomepageCA",
+    "ApplicationFR",
+    "HomepageFR",
+    "ApplicationDE",
+    "HomepageDE",
+    "ApplicationSU",
+    "HomepageSU",
+    "ApplicationUK",
+    "HomepageUK",
+    "ApplicationUS",
+    "HomepageUS",
+]
