+++ src/pages/germany/homepage.py
@@ -1,10 +1,9 @@
 from selenium.webdriver.remote.webdriver import WebDriver
 
-from src.pages.base import BasePage
+from src.pages.base_page import BasePage
 from src.config import Config as cfg
 from src.elements.homepage import (
     CookiesAcceptButton,
-    CloseLightboxButton,
     NavbarLoginButton,
     SignupTabButton,
     LoginTabButton,
@@ -17,8 +16,7 @@ from src.elements.homepage import (
     SignupLastNameField,
     SignupBirthDateField,
     SignupEmailField,
-    SignupFamilyRelField,
-    SignupFatherOption,
+    SignupGenderSelection,
     SignupPhoneField,
     SignupSubmitButton,
     ScrollDownTermsButton,
@@ -26,29 +24,27 @@ from src.elements.homepage import (
 )
 
 
-class HomepageUS(BasePage):
+class Homepage(BasePage):
     """Class containing actions available
-    on United States homepage
+    on Germany homepage
     """
 
     def __init__(self, driver: WebDriver) -> None:
         super().__init__(driver)
-
         # URL
         if cfg.ENVIRONMENT == "PROD":
-            self.base_url = "http://www.birthrightisrael.com/"
+            self.base_url = "https://de.birthrightisrael.com"
         else:
-            self.base_url = "http://www.taglit.info/"
+            self.base_url = "https://de.taglit.info"
 
         # Country
-        self.iso = "US"
+        self.iso = "DE"
 
         # Title
-        self.title = "A Free Trip to Israel | Taglit - Birthright Israel"
+        self.title = "Taglit - Birthright Israel"
 
         # Elements
         self.btn_cookies_cta = CookiesAcceptButton(driver)
-        self.btn_close_lightbox = CloseLightboxButton(driver)
         self.btn_navbar_login = NavbarLoginButton(driver)
         self.btn_signup_tab = SignupTabButton(driver)
         self.btn_login_tab = LoginTabButton(driver)
@@ -61,8 +57,7 @@ class HomepageUS(BasePage):
         self.fld_signup_lname = SignupLastNameField(driver)
         self.fld_signup_birth_date = SignupBirthDateField(driver)
         self.fld_signup_email = SignupEmailField(driver)
-        self.fld_family_rel = SignupFamilyRelField(driver)
-        self.rad_family_rel_father = SignupFatherOption(driver)
+        self.fld_signup_gender = SignupGenderSelection(driver)
         self.fld_signup_phone = SignupPhoneField(driver)
         self.btn_signup_button = SignupSubmitButton(driver)
         self.btn_signup_scroll_terms = ScrollDownTermsButton(driver)
@@ -85,12 +80,6 @@ class HomepageUS(BasePage):
             self.btn_cookies_cta.click()
         except Exception:
             raise Exception("failed to click cookies cta")
-
-    def close_lightbox(self):
-        try:
-            self.btn_close_lightbox.click()
-        except Exception:
-            raise Exception("failed to close lightbox")
 
     def open_login_popup(self):
         try:
@@ -164,17 +153,11 @@ class HomepageUS(BasePage):
         except Exception:
             raise Exception("failed to fill signup email")
 
-    def toggle_family_relation(self):
+    def fill_signup_gender(self, value):
         try:
-            self.fld_family_rel.click()
+            self.fld_signup_gender.select(value)
         except Exception:
-            raise Exception("failed to toggle family relation")
-
-    def check_family_relation_father(self):
-        try:
-            self.rad_family_rel_father.click()
-        except Exception:
-            raise Exception("failed to check father relation")
+            raise Exception("failed to fill signup gender")
 
     def fill_signup_phone(self, value):
         try:
