+++ src/pages/russia/__init__.py
@@ -1,0 +1,8 @@
+from .application import Application  # noqa
+from .homepage import Homepage  # noqa
+
+
+__all__=[
+    "Application",
+    "Homepage"
+]
