+++ src/pages/united_kingdom/__init__.py
@@ -1,0 +1,8 @@
+from .application import Application  # noqa
+from .homepage import Homepage  # noqa
+
+
+__all__=[
+    "Application",
+    "Homepage"
+]
